refactor: move speed bench report to analysis package
